requirements.txt
absl-py==2.1.0
altgraph==0.17.4
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.13.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
branca==0.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==2.0.1
ExifRead==3.0.0
fastjsonschema==2.19.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.48.1
fqdn==1.5.1
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.66.2
h5py==3.12.1
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.4.9
ipykernel==6.29.2
ipython==8.21.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
moviepy==1.0.3
mysqlclient==2.2.4
namex==0.0.8
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.0.8
notebook_shim==0.2.3
numexpr==2.10.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.0
overrides==7.7.0
packaging==20.9
pandas==2.2.0
pandasql==0.7.3
pandastable==0.13.1
pandocfilters==1.5.1
parso==0.8.3
pefile==2023.2.7
Pillow==10.1.0
platformdirs==4.2.0
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.8
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pynput==1.7.7
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytube==15.0.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.17.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.12.0
tensorflow-intel==2.17.0
tensorflowjs==3.18.0
termcolor==2.5.0
terminado==0.18.0
tinycss2==1.2.1
tkcalendar==1.6.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.0
vboxapi==1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xlrd==2.0.1
xyzservices==2024.4.0
